agent!: make cumulative execution quotas host-controlled
Represent unlimited, finite, and zero quotas explicitly; separate grants from allocation accounting and usage. Remove implicit cumulative work, strategy, process-count, and snapshot-size defaults while preserving capacity and cancellation controls.

Replace the snapshot contract, widen execution sequences, guard identity exhaustion, and validate unlimited child grants and recovery without compatibility paths.

+++ agent/examples/evaluator_optimizer/main_test.go
@@ -115,7 +115,7 @@ func TestConfigurationIsExplicitAndFinite(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			if _, _, err := newEvaluatorOptimizer(test.scores, test.threshold, test.maxIterations); err == nil {
+			if _, _, err := newEvaluatorOptimizer(test.scores, test.threshold, uint64(test.maxIterations)); err == nil {
 				t.Fatal("invalid evaluator-optimizer configuration was accepted")
 			}
 		})
